+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 NOTES:
 
-* The `morpheus-sdk` dependcy has been upgraded to version 0.2.7.
+* The `morpheus-sdk` dependcy has been upgraded to version 0.2.8.
 * Fix the `morpheus_provisioning_workflow` resource to properly support the "all" platform setting options.
 
 FEATURES:
 
 * **New Resource:** `morpheus_api_option_list`
+* **New Resource:** `morpheus_radio_list_option_type`
 
 ## 0.7.0 (December 16, 2022)